+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+
+### Fixed
+
+- Debounced event handlers are now cleared on unmount, resolving
+  [#231 Possible Regression](https://github.com/vue-leaflet/vue-leaflet/issues/231).
+
+
 ## [0.8.1] - 2023-01-29
 
 ### Fixed